#95 [add] near store option
