+++ app/src/main/res/drawable/ic_near_store.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="18dp"
+    android:height="19dp"
+    android:viewportWidth="18"
+    android:viewportHeight="19">
+  <group>
+    <clip-path
+        android:pathData="M0,0.5h18v18h-18z"/>
+    <path
+        android:pathData="M15.45,8.75V14.656C15.45,15.019 15.325,15.325 15.075,15.575C14.825,15.825 14.525,15.95 14.175,15.95H3.843C3.481,15.95 3.175,15.825 2.925,15.575C2.675,15.325 2.55,15.019 2.55,14.656V8.731C2.237,8.506 2.009,8.191 1.866,7.784C1.722,7.378 1.725,6.95 1.875,6.5L2.606,4.063C2.706,3.763 2.875,3.519 3.112,3.331C3.35,3.144 3.631,3.05 3.956,3.05H14.025C14.35,3.05 14.631,3.138 14.868,3.313C15.106,3.487 15.281,3.731 15.393,4.044L16.143,6.5C16.293,6.95 16.293,7.378 16.143,7.784C15.993,8.191 15.762,8.512 15.45,8.75ZM10.631,8.225C11.056,8.225 11.368,8.1 11.568,7.85C11.768,7.6 11.85,7.331 11.812,7.044L11.362,4.1H9.525V7.044C9.525,7.356 9.634,7.631 9.853,7.869C10.072,8.106 10.331,8.225 10.631,8.225ZM7.293,8.225C7.643,8.225 7.928,8.106 8.147,7.869C8.365,7.631 8.475,7.356 8.475,7.044V4.1H6.637L6.187,7.081C6.15,7.344 6.234,7.6 6.441,7.85C6.647,8.1 6.931,8.225 7.293,8.225ZM4.012,8.225C4.3,8.225 4.543,8.128 4.743,7.934C4.943,7.74 5.068,7.494 5.118,7.194L5.568,4.1H3.918C3.843,4.1 3.781,4.115 3.731,4.147C3.681,4.178 3.643,4.231 3.618,4.306L2.887,6.725C2.775,7.088 2.821,7.428 3.027,7.747C3.234,8.066 3.562,8.225 4.012,8.225ZM14.006,8.225C14.406,8.225 14.722,8.072 14.953,7.766C15.184,7.459 15.237,7.113 15.112,6.725L14.362,4.287C14.337,4.225 14.3,4.178 14.25,4.147C14.2,4.115 14.137,4.1 14.062,4.1H12.45L12.9,7.194C12.95,7.494 13.075,7.74 13.275,7.934C13.475,8.128 13.718,8.225 14.006,8.225Z"
+        android:fillColor="#666666"/>
+  </group>
+</vector>
