+++ app/src/main/res/values/strings.xml
@@ -142,6 +142,8 @@
     <string name="friend_restaurant">친구 음식점</string>
     <string name="nearby_restaurant">가까운 음식점</string>
     <string name="check_the_real_location">현 위치에서 검색</string>
+    <string name="not_find_restaurant">음식점을 찾을 수 없습니다.</string>
+    <string name="choose_restaurant_options">찾는 음식점 옵션을 선택해주세요.</string>
 
     <string name="move_setting">설정으로 이동</string>
     <string name="location_service_term_text">위치 서비스를 사용할 수 없습니다. 기기의 \'설정\'에서 위치 서비스를 켜주세요.(필수권한)</string>
