+++ domain/src/main/java/com/jjbaksa/domain/model/mainpage/JjCategory.kt
@@ -1,5 +1,5 @@
 package com.jjbaksa.domain.model.mainpage
 
 enum class JjCategory {
-    FRIEND, BOOKMARK
+    NEAR_STORE, FRIEND, BOOKMARK
 }
